requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
geoalchemy2
python-dotenv
reportlab
matplotlib
geopandas
Pillow
contextily
ezdxf